+++ ORStools/ORStoolsPlugin.py
@@ -27,8 +27,9 @@
  ***************************************************************************/
 """
 
+from qgis.gui import QgisInterface
 from qgis.core import QgsApplication, QgsSettings
-from PyQt5.QtCore import QTranslator, qVersion, QCoreApplication
+from qgis.PyQt.QtCore import QTranslator, qVersion, QCoreApplication, QLocale
 import os.path
 
 from .gui import ORStoolsDialog
@@ -40,7 +41,7 @@ class ORStools:
 
     # noinspection PyTypeChecker,PyArgumentList,PyCallByClass
 
-    def __init__(self, iface):
+    def __init__(self, iface: QgisInterface) -> None:
         """Constructor.
 
         :param iface: An interface instance that will be passed to this class
@@ -55,25 +56,32 @@ class ORStools:
         self.plugin_dir = os.path.dirname(__file__)
 
         # initialize locale
-        locale = QgsSettings().value("locale/userLocale")[0:2]
-        locale_path = os.path.join(self.plugin_dir, "i18n", "orstools_{}.qm".format(locale))
+        try:
+            locale = QgsSettings().value("locale/userLocale")
+            if not locale:
+                locale = QLocale().name()
+            locale = locale[0:2]
 
-        if os.path.exists(locale_path):
-            self.translator = QTranslator()
-            self.translator.load(locale_path)
+            locale_path = os.path.join(self.plugin_dir, "i18n", "orstools_{}.qm".format(locale))
 
-            if qVersion() > "4.3.3":
-                QCoreApplication.installTranslator(self.translator)
+            if os.path.exists(locale_path):
+                self.translator = QTranslator()
+                self.translator.load(locale_path)
 
+                if qVersion() > "4.3.3":
+                    QCoreApplication.installTranslator(self.translator)
+        except TypeError:
+            pass
+          
         self.add_default_provider_to_settings()
 
-    def initGui(self):
+    def initGui(self) -> None:
         """Create the menu entries and toolbar icons inside the QGIS GUI."""
 
         QgsApplication.processingRegistry().addProvider(self.provider)
         self.dialog.initGui()
 
-    def unload(self):
+    def unload(self) -> None:
         """remove menu entry and toolbar icons"""
         QgsApplication.processingRegistry().removeProvider(self.provider)
         self.dialog.unload()