+++ ORStools/proc/directions_points_layers_proc.py
@@ -27,6 +27,9 @@
  ***************************************************************************/
 """
 
+from typing import Dict
+
+from qgis._core import QgsField
 from qgis.core import (
     QgsWkbTypes,
     QgsCoordinateReferenceSystem,
@@ -34,9 +37,14 @@ from qgis.core import (
     QgsProcessingParameterField,
     QgsProcessingParameterFeatureSource,
     QgsProcessingParameterEnum,
+    QgsProcessingParameterNumber,
+    QgsProcessingParameterString,
+    QgsProcessingFeatureSource,
+    QgsProcessingContext,
+    QgsProcessingFeedback,
 )
 
-from ORStools.common import directions_core, PROFILES, PREFERENCES
+from ORStools.common import directions_core, PROFILES, PREFERENCES, EXTRA_INFOS
 from ORStools.utils import transform, exceptions, logger
 from .base_processing_algorithm import ORSBaseProcessingAlgorithm
 
@@ -45,22 +53,25 @@ from .base_processing_algorithm import ORSBaseProcessingAlgorithm
 class ORSDirectionsPointsLayersAlgo(ORSBaseProcessingAlgorithm):
     def __init__(self):
         super().__init__()
-        self.ALGO_NAME = "directions_from_points_2_layers"
-        self.GROUP = "Directions"
+        self.ALGO_NAME: str = "directions_from_points_2_layers"
+        self.GROUP: str = "Directions"
         self.MODE_SELECTION: list = ["Row-by-Row", "All-by-All"]
-        self.IN_START = "INPUT_START_LAYER"
-        self.IN_START_FIELD = "INPUT_START_FIELD"
-        self.IN_SORT_START_BY = "INPUT_SORT_START_BY"
-        self.IN_END = "INPUT_END_LAYER"
-        self.IN_END_FIELD = "INPUT_END_FIELD"
-        self.IN_SORT_END_BY = "INPUT_SORT_END_BY"
-        self.IN_PREFERENCE = "INPUT_PREFERENCE"
-        self.IN_MODE = "INPUT_MODE"
-        self.PARAMETERS = [
+        self.IN_START: str = "INPUT_START_LAYER"
+        self.IN_START_FIELD: str = "INPUT_START_FIELD"
+        self.IN_SORT_START_BY: str = "INPUT_SORT_START_BY"
+        self.IN_END: str = "INPUT_END_LAYER"
+        self.IN_END_FIELD: str = "INPUT_END_FIELD"
+        self.IN_SORT_END_BY: str = "INPUT_SORT_END_BY"
+        self.IN_PREFERENCE: str = "INPUT_PREFERENCE"
+        self.IN_MODE: str = "INPUT_MODE"
+        self.EXTRA_INFO: str = "EXTRA_INFO"
+        self.CSV_FACTOR: str = "CSV_FACTOR"
+        self.CSV_COLUMN: str = "CSV_COLUMN"
+        self.PARAMETERS: list = [
             QgsProcessingParameterFeatureSource(
                 name=self.IN_START,
                 description=self.tr("Input Start Point layer"),
-                types=[QgsProcessing.TypeVectorPoint],
+                types=[QgsProcessing.SourceType.TypeVectorPoint],
             ),
             QgsProcessingParameterField(
                 name=self.IN_START_FIELD,
@@ -79,7 +90,7 @@ class ORSDirectionsPointsLayersAlgo(ORSBaseProcessingAlgorithm):
             QgsProcessingParameterFeatureSource(
                 name=self.IN_END,
                 description=self.tr("Input End Point layer"),
-                types=[QgsProcessing.TypeVectorPoint],
+                types=[QgsProcessing.SourceType.TypeVectorPoint],
             ),
             QgsProcessingParameterField(
                 name=self.IN_END_FIELD,
@@ -107,11 +118,34 @@ class ORSDirectionsPointsLayersAlgo(ORSBaseProcessingAlgorithm):
                 self.MODE_SELECTION,
                 defaultValue=self.MODE_SELECTION[0],
             ),
+            QgsProcessingParameterEnum(
+                self.EXTRA_INFO,
+                self.tr("Extra Info"),
+                options=EXTRA_INFOS,
+                allowMultiple=True,
+                optional=True,
+            ),
+            QgsProcessingParameterNumber(
+                self.CSV_FACTOR,
+                self.tr("Csv Factor (needs Csv Column and csv in Extra Info)"),
+                type=QgsProcessingParameterNumber.Double,
+                minValue=0,
+                maxValue=1,
+                defaultValue=None,
+                optional=True,
+            ),
+            QgsProcessingParameterString(
+                self.CSV_COLUMN,
+                self.tr("Csv Column (needs Csv Factor and csv in Extra Info)"),
+                optional=True,
+            ),
         ]
 
     # TODO: preprocess parameters to options the range cleanup below:
     # https://www.qgis.org/pyqgis/master/core/Processing/QgsProcessingAlgorithm.html#qgis.core.QgsProcessingAlgorithm.preprocessParameters
-    def processAlgorithm(self, parameters, context, feedback):
+    def processAlgorithm(
+        self, parameters: dict, context: QgsProcessingContext, feedback: QgsProcessingFeedback
+    ) -> Dict[str, str]:
         ors_client = self._get_ors_client_from_provider(parameters[self.IN_PROVIDER], feedback)
 
         profile = dict(enumerate(PROFILES))[parameters[self.IN_PROFILE]]
@@ -121,6 +155,16 @@ class ORSDirectionsPointsLayersAlgo(ORSBaseProcessingAlgorithm):
         mode = dict(enumerate(self.MODE_SELECTION))[parameters[self.IN_MODE]]
 
         options = self.parseOptions(parameters, context)
+
+        csv_factor = self.parameterAsDouble(parameters, self.CSV_FACTOR, context)
+        csv_column = self.parameterAsString(parameters, self.CSV_COLUMN, context)
+        if csv_factor > 0:
+            options["profile_params"] = {
+                "weightings": {"csv_factor": csv_factor, "csv_column": csv_column}
+            }
+
+        extra_info = self.parameterAsEnums(parameters, self.EXTRA_INFO, context)
+        extra_info = [EXTRA_INFOS[i] for i in extra_info]
 
         # Get parameter values
         source = self.parameterAsSource(parameters, self.IN_START, context)
@@ -168,14 +212,16 @@ class ORSDirectionsPointsLayersAlgo(ORSBaseProcessingAlgorithm):
             field_types.update({"from_type": source_field.type()})
         if destination_field:
             field_types.update({"to_type": destination_field.type()})
-        sink_fields = directions_core.get_fields(**field_types)
+        sink_fields = directions_core.get_fields(
+            **field_types, extra_info=extra_info, two_layers=True
+        )
 
         (sink, dest_id) = self.parameterAsSink(
             parameters,
             self.OUT,
             context,
             sink_fields,
-            QgsWkbTypes.LineString,
+            QgsWkbTypes.Type.LineString,
             QgsCoordinateReferenceSystem.fromEpsgId(4326),
         )
 
@@ -186,7 +232,7 @@ class ORSDirectionsPointsLayersAlgo(ORSBaseProcessingAlgorithm):
                 break
 
             params = directions_core.build_default_parameters(
-                preference, coordinates=coordinates, options=options
+                preference, coordinates=coordinates, options=options, extra_info=extra_info
             )
 
             try:
@@ -199,11 +245,18 @@ class ORSDirectionsPointsLayersAlgo(ORSBaseProcessingAlgorithm):
                 logger.log(msg)
                 continue
 
-            sink.addFeature(
-                directions_core.get_output_feature_directions(
-                    response, profile, preference, from_value=values[0], to_value=values[1]
+            if extra_info:
+                feats = directions_core.get_extra_info_features_directions(
+                    response, extra_info, values
                 )
-            )
+                for feat in feats:
+                    sink.addFeature(feat)
+            else:
+                sink.addFeature(
+                    directions_core.get_output_feature_directions(
+                        response, profile, preference, from_value=values[0], to_value=values[1]
+                    )
+                )
 
             counter += 1
             feedback.setProgress(int(100.0 / route_count * counter))
@@ -211,18 +264,25 @@ class ORSDirectionsPointsLayersAlgo(ORSBaseProcessingAlgorithm):
         return {self.OUT: dest_id}
 
     @staticmethod
-    def _get_route_dict(source, source_field, sort_start, destination, destination_field, sort_end):
+    def _get_route_dict(
+        source: QgsProcessingFeatureSource,
+        source_field: QgsField,
+        sort_start,
+        destination: QgsProcessingFeatureSource,
+        destination_field: QgsField,
+        sort_end,
+    ) -> dict:
         """
         Compute route_dict from input layer.
 
         :param source: Input from layer
-        :type source: QgsProcessingParameterFeatureSource
+        :type source: QgsProcessingFeatureSource
 
         :param source_field: ID field from layer.
         :type source_field: QgsField
 
         :param destination: Input to layer.
-        :type destination: QgsProcessingParameterFeatureSource
+        :type destination: QgsProcessingFeatureSource
 
         :param destination_field: ID field to layer.
         :type destination_field: QgsField