+++ ORStools/proc/matrix_proc.py
@@ -27,6 +27,8 @@
  ***************************************************************************/
 """
 
+from typing import Dict
+
 from qgis.core import (
     QgsWkbTypes,
     QgsFeature,
@@ -36,9 +38,11 @@ from qgis.core import (
     QgsProcessingException,
     QgsProcessingParameterField,
     QgsProcessingParameterFeatureSource,
+    QgsProcessingContext,
+    QgsProcessingFeedback,
 )
 
-from PyQt5.QtCore import QVariant
+from qgis.PyQt.QtCore import QVariant
 
 from ORStools.common import PROFILES
 from ORStools.utils import transform, exceptions, logger
@@ -49,17 +53,17 @@ from .base_processing_algorithm import ORSBaseProcessingAlgorithm
 class ORSMatrixAlgo(ORSBaseProcessingAlgorithm):
     def __init__(self):
         super().__init__()
-        self.ALGO_NAME = "matrix_from_layers"
-        self.GROUP = "Matrix"
-        self.IN_START = "INPUT_START_LAYER"
-        self.IN_START_FIELD = "INPUT_START_FIELD"
-        self.IN_END = "INPUT_END_LAYER"
-        self.IN_END_FIELD = "INPUT_END_FIELD"
-        self.PARAMETERS = [
+        self.ALGO_NAME: str = "matrix_from_layers"
+        self.GROUP: str = "Matrix"
+        self.IN_START: str = "INPUT_START_LAYER"
+        self.IN_START_FIELD: str = "INPUT_START_FIELD"
+        self.IN_END: str = "INPUT_END_LAYER"
+        self.IN_END_FIELD: str = "INPUT_END_FIELD"
+        self.PARAMETERS: list = [
             QgsProcessingParameterFeatureSource(
                 name=self.IN_START,
                 description=self.tr("Input Start Point layer"),
-                types=[QgsProcessing.TypeVectorPoint],
+                types=[QgsProcessing.SourceType.TypeVectorPoint],
             ),
             QgsProcessingParameterField(
                 name=self.IN_START_FIELD,
@@ -71,7 +75,7 @@ class ORSMatrixAlgo(ORSBaseProcessingAlgorithm):
             QgsProcessingParameterFeatureSource(
                 name=self.IN_END,
                 description=self.tr("Input End Point layer"),
-                types=[QgsProcessing.TypeVectorPoint],
+                types=[QgsProcessing.SourceType.TypeVectorPoint],
             ),
             QgsProcessingParameterField(
                 name=self.IN_END_FIELD,
@@ -82,7 +86,9 @@ class ORSMatrixAlgo(ORSBaseProcessingAlgorithm):
             ),
         ]
 
-    def processAlgorithm(self, parameters, context, feedback):
+    def processAlgorithm(
+        self, parameters: dict, context: QgsProcessingContext, feedback: QgsProcessingFeedback
+    ) -> Dict[str, str]:
         ors_client = self._get_ors_client_from_provider(parameters[self.IN_PROVIDER], feedback)
 
         # Get profile value
@@ -106,7 +112,7 @@ class ORSMatrixAlgo(ORSBaseProcessingAlgorithm):
         # Abort when MultiPoint type
         if (
             QgsWkbTypes.flatType(source.wkbType()) or QgsWkbTypes.flatType(destination.wkbType())
-        ) == QgsWkbTypes.MultiPoint:
+        ) == QgsWkbTypes.Type.MultiPoint:
             raise QgsProcessingException(
                 "TypeError: Multipoint Layers are not accepted. Please convert to single geometry layer."
             )
@@ -177,7 +183,7 @@ class ORSMatrixAlgo(ORSBaseProcessingAlgorithm):
             logger.log(msg)
 
         (sink, dest_id) = self.parameterAsSink(
-            parameters, self.OUT, context, sink_fields, QgsWkbTypes.NoGeometry
+            parameters, self.OUT, context, sink_fields, QgsWkbTypes.Type.NoGeometry
         )
 
         sources_attributes = [