+++ CHANGELOG.md
@@ -41,15 +41,55 @@ RELEASING:
  -->
 
 ## Unreleased
+
+### Changed
 - Use QgsSettings instead of config.yml file to avoid deletion of providers on update ([#108](https://github.com/GIScience/orstools-qgis-plugin/issues/108))
+
+## [1.8.4] - 2024-07-29
+
+### Fixed
+- issue with missing locale value for non-default user([#271](https://github.com/GIScience/orstools-qgis-plugin/issues/271))
+
+## [1.8.3] - 2024-05-29
+
+### Fixed
+- Add csv\_column parameter to request made by points\_layers\_proc([#260](https://github.com/GIScience/orstools-qgis-plugin/issues/260))
+- make extra\_info work with two points layers
+
+## [1.8.2] - 2024-05-20
+
+### Fixed
+- use python 3.9-compatible type hint
+
+## [1.8.1] - 2024-05-17
+
+### Fixed
+- use if-else instead of structural pattern matching
+
+## [1.8.0] - 2024-05-17
+
+### Added
+- Add support for decimal ranges with isochrones([#237](https://github.com/GIScience/orstools-qgis-plugin/issues/237))
+- Add hint for joining with `Layer ID Field` ([#143](https://github.com/GIScience/orstools-qgis-plugin/issues/143))
+- Add option to export order of optimization route points ([#145](https://github.com/GIScience/orstools-qgis-plugin/issues/145))
+- Add `extra_info` parameter to directions processing algorithms ([#251](https://github.com/GIScience/orstools-qgis-plugin/pull/251))
+
+### Changed
+- Rename `Ok` button in configuration window to `Save` for clarification ([#241](https://github.com/GIScience/orstools-qgis-plugin/issues/241))
+- Replace PyQt5 imports with qgis.PyQt imports to prepare for Qt6-QGis builds
 
 ### Fixed
 - QGis crashes when selecting more than two vertices for deletion ([#230](https://github.com/GIScience/orstools-qgis-plugin/issues/230))
 - Vertices on canvas not depicted fully with n having more than one digit in length ([#235](https://github.com/GIScience/orstools-qgis-plugin/issues/235))
 - Replace qt QSettings with QgsSettings for centralized configuration management ([#239](https://github.com/GIScience/orstools-qgis-plugin/issues/239))
+- Point Annotations stay after saving project and not deleting them manually ([#229](https://github.com/GIScience/orstools-qgis-plugin/issues/229))
+- Issue with MultiPoint-layers and optimization
+- Improved type hints
 
+
+# Unreleased
 ### Added
-- Add support for decimal ranges with isochrones([#237](https://github.com/GIScience/orstools-qgis-plugin/issues/237))
+- Unit- and e2e-testing
 
 ## [1.7.1] - 2024-01-15
 
@@ -231,7 +271,12 @@ RELEASING:
 - first working version of ORS Tools, after replacing OSM Tools plugin
 
 
-[unreleased]: https://github.com/GIScience/orstools-qgis-plugin/compare/v1.7.1...HEAD
+[unreleased]: https://github.com/GIScience/orstools-qgis-plugin/compare/v1.8.4...HEAD
+[1.8.4]: https://github.com/GIScience/orstools-qgis-plugin/compare/v1.8.3...v1.8.4
+[1.8.3]: https://github.com/GIScience/orstools-qgis-plugin/compare/v1.8.2...v1.8.3
+[1.8.2]: https://github.com/GIScience/orstools-qgis-plugin/compare/v1.8.1...v1.8.2
+[1.8.1]: https://github.com/GIScience/orstools-qgis-plugin/compare/v1.8.0...v1.8.1
+[1.8.0]: https://github.com/GIScience/orstools-qgis-plugin/compare/v1.7.1...v1.8.0
 [1.7.1]: https://github.com/GIScience/orstools-qgis-plugin/compare/v1.7.0...v1.7.1
 [1.7.0]: https://github.com/GIScience/orstools-qgis-plugin/compare/v1.6.0...v1.7.0
 [1.6.0]: https://github.com/GIScience/orstools-qgis-plugin/compare/v1.5.3...v1.6.0